requirements.txt
PyMuPDF>=1.23.0
pillow>=10.0.0